activate gh actions (#4)
* enable gh actions

* install deps

* fix

* fix macos

* fix macos

* windohs

* cmake eigen3 dir

* check path

* install eigen

* urgh

* donkey work

* boost again

* simplify"

* can't have nie things

* other dir

* choco eigen

* whaat

* link

* M_PI

* grr

* with VS config

* double nan

* that's what PhDs are for

* yep

* dumdidum

+++ include/kde1d/tools.hpp
@@ -30,7 +30,7 @@ Eigen::MatrixXd
 unaryExpr_or_nan_int(const Eigen::MatrixXi& x, const T& func)
 {
   return x.unaryExpr([&func](int y) {
-    if (std::isnan(y)) {
+    if (std::isnan(static_cast<double>(y))) {
       return std::numeric_limits<double>::quiet_NaN();
     } else {
       return func(y);
